refactor(infrastructure): streamline dependency injection and server initialization
This commit focuses on simplifying the server initialization process and improving dependency injection:

- Simplified main.py server startup logic
- Removed redundant server creation functions
- Improved tool registration and service container initialization
- Cleaned up imports and removed commented-out code
- Updated file structure for SQL-related models and validators
- Removed unnecessary logging statements
- Adjusted SDK and API client methods to be more straightforward
