+++ CHANGELOG.MD
@@ -28,7 +28,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
   - return a specific path and method
   - return all paths and emthods
 - Added code coverage
-
+9. Infra:
+- Streamlined dependency injection for easier testing
 
 ## Changed
 - Updated README.md with detailed feature description.
